refactor promoted builds support into 9a86fdb753
