+++ job-dsl-core/src/main/groovy/javaposse/jobdsl/dsl/helpers/promotions/ConditionsContext.groovy
@@ -1,8 +1,8 @@
 package javaposse.jobdsl.dsl.helpers.promotions
 
 import javaposse.jobdsl.dsl.helpers.BuildParametersContext
-import javaposse.jobdsl.dsl.helpers.Context
-import javaposse.jobdsl.dsl.helpers.ContextHelper
+import javaposse.jobdsl.dsl.Context
+import javaposse.jobdsl.dsl.ContextHelper
 
 class ConditionsContext implements Context {
